feat: update loading artwork indicator
